+++ stdlib/Serialization/src/Serialization.jl
@@ -694,6 +694,16 @@ function serialize(s::IO, x)
     serialize(ss, x)
 end
 
+"""
+    serialize(filename::AbstractString, value)
+
+Open a file and serialize the given value to it.
+
+!!! compat "Julia 1.1"
+    This method is available as of Julia 1.1.
+"""
+serialize(filename::AbstractString, x) = open(io->serialize(io, x), filename, "w")
+
 ## deserializing values ##
 
 """
@@ -706,6 +716,16 @@ the data read. Malformed data can result in process termination. The caller has 
 the integrity and correctness of data read from `stream`.
 """
 deserialize(s::IO) = deserialize(Serializer(s))
+
+"""
+    deserialize(filename::AbstractString)
+
+Open a file and deserialize its contents.
+
+!!! compat "Julia 1.1"
+    This method is available as of Julia 1.1.
+"""
+deserialize(filename::AbstractString) = open(deserialize, filename)
 
 function deserialize(s::AbstractSerializer)
     handle_deserialize(s, Int32(read(s.io, UInt8)::UInt8))